error パッケージから exceptions パッケージにリネームする
