+++ src/main/kotlin/micronaut/bookman/controller/ErrorResponseBody.kt
@@ -1,5 +1,5 @@
 package micronaut.bookman.controller
 
-import micronaut.bookman.error.ErrorCode
+import micronaut.bookman.exceptions.ErrorCode
 
 data class ErrorResponseBody(val id: ErrorCode, val message: String)
