+++ src/main/kotlin/micronaut/bookman/controller/person/PersonsController.kt
@@ -2,9 +2,9 @@ package micronaut.bookman.controller.person
 
 import io.micronaut.http.HttpResponse
 import io.micronaut.http.annotation.Controller
-import micronaut.bookman.error.PersonErrorResponseSyntax.toResponseBody
+import micronaut.bookman.exceptions.PersonErrorResponseSyntax.toResponseBody
 import micronaut.bookman.controller.UnitResponse
-import micronaut.bookman.domain.person.error.NoPersonException
+import micronaut.bookman.domain.person.exceptions.NoPersonException
 import micronaut.bookman.query.PersonSearchQueryService
 import micronaut.bookman.usecase.LibrarianPersonUseCase
