+++ src/main/kotlin/micronaut/bookman/domain/book/exceptions/NoBookException.kt
@@ -1,4 +1,4 @@
-package micronaut.bookman.domain.book.error
+package micronaut.bookman.domain.book.exceptions
 
 class NoBookException private constructor(
         override val message: String?,
