+++ src/main/kotlin/micronaut/bookman/domain/person/exceptions/DuplicatePersonException.kt
@@ -1,9 +1,8 @@
-package micronaut.bookman.domain.person.error
+package micronaut.bookman.domain.person.exceptions
 
-class NoPersonException private constructor(
+class DuplicatePersonException private constructor(
         override val message: String?,
         override val cause: Throwable?
 ) : PersonDomainException(message, cause) {
-    constructor(any: List<String>): this("Person(any = ${any})")
     constructor(id: String) : this("Person(id = ${id})", null)
 }
