+++ src/main/kotlin/micronaut/bookman/domain/person/exceptions/NoPersonException.kt
@@ -1,8 +1,9 @@
-package micronaut.bookman.domain.person.error
+package micronaut.bookman.domain.person.exceptions
 
-class DuplicatePersonException private constructor(
+class NoPersonException private constructor(
         override val message: String?,
         override val cause: Throwable?
 ) : PersonDomainException(message, cause) {
+    constructor(any: List<String>): this("Person(any = ${any})")
     constructor(id: String) : this("Person(id = ${id})", null)
 }
