+++ src/main/kotlin/micronaut/bookman/exceptions/BookErrorResponseSyntax.kt
@@ -1,7 +1,7 @@
-package micronaut.bookman.error
+package micronaut.bookman.exceptions
 
 import micronaut.bookman.controller.ErrorResponseBody
-import micronaut.bookman.domain.book.error.NoBookException
+import micronaut.bookman.domain.book.exceptions.NoBookException
 
 object BookErrorResponseSyntax {
     fun NoBookException.toResponseBody() = ErrorResponseBody(ErrorCode.BOOK_NOT_FOUND, this.message
