+++ src/main/kotlin/micronaut/bookman/exceptions/PersonErrorResponseSyntax.kt
@@ -1,7 +1,7 @@
-package micronaut.bookman.error
+package micronaut.bookman.exceptions
 
 import micronaut.bookman.controller.ErrorResponseBody
-import micronaut.bookman.domain.person.error.NoPersonException
+import micronaut.bookman.domain.person.exceptions.NoPersonException
 
 object PersonErrorResponseSyntax {
     fun NoPersonException.toResponseBody() = ErrorResponseBody(ErrorCode.PERSON_NOT_FOUND, this.message
