+++ src/main/kotlin/micronaut/bookman/infra/book/DBBookRepository.kt
@@ -4,14 +4,14 @@ import io.micronaut.context.annotation.Primary
 import micronaut.bookman.domain.book.Book
 import micronaut.bookman.domain.book.BookAuthor
 import micronaut.bookman.domain.book.BookRepository
-import micronaut.bookman.domain.book.error.DuplicateBookException
-import micronaut.bookman.domain.book.error.NoBookException
-import micronaut.bookman.domain.person.error.NoPersonException
+import micronaut.bookman.domain.book.exceptions.DuplicateBookException
+import micronaut.bookman.domain.book.exceptions.NoBookException
+import micronaut.bookman.domain.person.exceptions.NoPersonException
 import micronaut.bookman.infra.schema.BookAuthorTable
 import micronaut.bookman.infra.schema.BookTable
 import micronaut.bookman.infra.DatabaseTrait
-import micronaut.bookman.infra.error.IllegalDatabaseSchema
-import micronaut.bookman.infra.error.InfraException
+import micronaut.bookman.infra.exceptions.IllegalDatabaseSchema
+import micronaut.bookman.infra.exceptions.InfraException
 import org.jetbrains.exposed.exceptions.ExposedSQLException
 import org.jetbrains.exposed.sql.*
 import org.jetbrains.exposed.sql.transactions.transaction
