+++ src/main/kotlin/micronaut/bookman/infra/exceptions/InfraException.kt
@@ -1,6 +1,6 @@
-package micronaut.bookman.infra.error
+package micronaut.bookman.infra.exceptions
 
-import micronaut.bookman.error.ApplicationException
+import micronaut.bookman.exceptions.ApplicationException
 
 open class InfraException(override val message: String?, override val cause: Throwable?) : ApplicationException(message, cause) {
     constructor(message: String?) : this(message, null)
