+++ src/test/kotlin/micronaut/bookman/usecase/LibrarianBookUseCaseTest.kt
@@ -8,10 +8,9 @@ import io.micronaut.test.annotation.MicronautTest
 import micronaut.bookman.SpecWithDataSource
 import micronaut.bookman.domain.book.Book
 import micronaut.bookman.domain.book.BookRepository
-import micronaut.bookman.domain.book.error.NoBookException
-import micronaut.bookman.domain.person.FullName
+import micronaut.bookman.domain.book.exceptions.NoBookException
 import micronaut.bookman.domain.person.PersonRepository
-import micronaut.bookman.domain.person.error.NoPersonException
+import micronaut.bookman.domain.person.exceptions.NoPersonException
 import java.util.*
 import javax.sql.DataSource
